refactor(app): shift around code to avoid import errors
