+++ package.json
@@ -18,6 +18,7 @@
     "eslint-config-next": "13.3.4",
     "flowbite": "^1.6.5",
     "flowbite-react": "^0.4.4",
+    "jotai": "^2.0.4",
     "next": "13.3.4",
     "openai": "^3.2.1",
     "postcss": "8.4.23",
@@ -28,6 +29,8 @@
     "typescript": "5.0.4"
   },
   "devDependencies": {
+    "@swc-jotai/debug-label": "^0.0.10",
+    "@swc-jotai/react-refresh": "^0.0.8",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.2.8"
   }
