+++ src/app/page.tsx
@@ -1,4 +1,5 @@
-import { InfoSection, UserInput } from '@/components/client';
+import { InfoSection } from '@/components/client/info-section';
+import { UserInput } from '@/components/client/user-input';
 
 export default async function Home() {
 	return (
