+++ src/app/providers.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { Provider } from 'jotai';
+
+export function Providers({ children }: { children: React.ReactNode }) {
+	return <Provider>{children}</Provider>;
+}
