--- src/components/client/index.ts
@@ -1,3 +1,0 @@
-export * from './info-section';
-export * from './user-input';
-export * from './nav-bar';
