+++ src/components/client/loading-spinner-text/index.ts
@@ -1,0 +1,1 @@
+export { LoadingSpinner } from './loading-spinner';
