+++ src/components/client/loading-spinner-text/loading-spinner.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Spinner } from 'flowbite-react';
 
 export const LoadingSpinner = () => {
