+++ src/components/server/app-layout/index.ts
@@ -1,0 +1,1 @@
+export { AppLayout } from './page-layout';
