--- src/components/server/index.ts
@@ -1,3 +1,0 @@
-export * from './loading-spinner-text';
-export * from './used-prompt';
-export * from './container';
