--- src/components/server/loading-spinner-text/index.ts
@@ -1,1 +1,0 @@
-export * from './loading-spinner';
